refactor(react/runtime): split `updateMainThread` out of `patchUpdate.ts`
