+++ packages/react/runtime/src/lifecycle/patch/updateMainThread.ts
@@ -1,0 +1,69 @@
+// Copyright 2025 The Lynx Authors. All rights reserved.
+// Licensed under the Apache License Version 2.0 that can be found in the
+// LICENSE file in the root directory of this source tree.
+
+import { clearDelayedWorklets, updateWorkletRefInitValueChanges } from '@lynx-js/react/worklet-runtime/bindings';
+
+import { snapshotPatchApply } from './snapshotPatchApply.js';
+import { LifecycleConstant } from '../../lifecycleConstant.js';
+import { __pendingListUpdates } from '../../list.js';
+import { PerformanceTimingKeys, markTiming, setPipeline } from '../../lynx/performance.js';
+import { takeGlobalRefPatchMap } from '../../snapshot/ref.js';
+import { __page } from '../../snapshot.js';
+import { isEmptyObject } from '../../utils.js';
+import { getReloadVersion } from '../pass.js';
+import type { Patch, PatchOptions } from './commit.js';
+
+function updateMainThread(
+  { data, patchOptions }: {
+    data: string;
+    patchOptions: PatchOptions;
+  },
+): void {
+  if ((patchOptions.reloadVersion ?? 0) < getReloadVersion()) {
+    return;
+  }
+
+  setPipeline(patchOptions.pipelineOptions);
+  markTiming(PerformanceTimingKeys.parse_changes_start);
+  let { snapshotPatch, workletRefInitValuePatch, flushOptions } = JSON.parse(data) as Patch;
+  markTiming(PerformanceTimingKeys.parse_changes_end);
+
+  markTiming(PerformanceTimingKeys.patch_changes_start);
+  updateWorkletRefInitValueChanges(workletRefInitValuePatch);
+  __pendingListUpdates.clear();
+  if (snapshotPatch) {
+    snapshotPatchApply(snapshotPatch);
+  }
+  __pendingListUpdates.flush();
+  // console.debug('********** Lepus updatePatch:');
+  // printSnapshotInstance(snapshotInstanceManager.values.get(-1)!);
+
+  commitMainThreadPatchUpdate(patchOptions.commitTaskId);
+  if (patchOptions.isHydration) {
+    clearDelayedWorklets();
+  }
+  markTiming(PerformanceTimingKeys.patch_changes_end);
+  flushOptions ||= {};
+  if (patchOptions.pipelineOptions) {
+    flushOptions.pipelineOptions = patchOptions.pipelineOptions;
+  }
+  // TODO: triggerDataUpdated?
+  __FlushElementTree(__page, flushOptions);
+}
+
+function injectUpdateMainThread(): void {
+  Object.assign(globalThis, { [LifecycleConstant.patchUpdate]: updateMainThread });
+}
+
+function commitMainThreadPatchUpdate(commitTaskId?: number): void {
+  const refPatch = takeGlobalRefPatchMap();
+  if (!isEmptyObject(refPatch)) {
+    __OnLifecycleEvent([LifecycleConstant.ref, { commitTaskId, refPatch: JSON.stringify(refPatch) }]);
+  }
+}
+
+/**
+ * @internal
+ */
+export { injectUpdateMainThread, commitMainThreadPatchUpdate };
