+++ packages/react/runtime/src/list.ts
@@ -2,7 +2,7 @@
 // Licensed under the Apache License Version 2.0 that can be found in the
 // LICENSE file in the root directory of this source tree.
 import { hydrate } from './hydrate.js';
-import { commitMainThreadPatchUpdate } from './lifecycle/patch/patchUpdate.js';
+import { commitMainThreadPatchUpdate } from './lifecycle/patch/updateMainThread.js';
 import type { SnapshotInstance } from './snapshot.js';
 
 export interface ListUpdateInfo {
